refactor(p2p): extract bearer readiness operations
